% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfun_boundary2alpha.R
\name{boundary2alpha}
\alias{boundary2alpha}
\title{Convert normal critical boundaries to cumulative alpha levels.}
\usage{
boundary2alpha(cvec, t)
}
\arguments{
\item{cvec}{a vector of critical boundaries}

\item{t}{a vector of information times}
}
\value{
alphas, a vector of cumulative errors
}
\description{
This function converts normal critical boundaries to cumulative alpha levels.
}
\examples{
t <- c(0.5,0.8,1)
iuse <- 4
result <- gbounds(t=t, iuse=iuse)
print(result)
boundary2alpha(cvec=result$bd, t=t)
}
\author{
Jiangtao Gou

Fengqing Zhang
}
