% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfun_corrMatGenerator.R
\name{corrMatGenerator}
\alias{corrMatGenerator}
\title{Correlation matrix generator}
\usage{
corrMatGenerator(tp, ts, rhops)
}
\arguments{
\item{tp}{an information fraction vector of Hp}

\item{ts}{an information fraction vector of Hs}

\item{rhops}{a number that shows the correlation coefficient between the test statistics of the primary endpoint and that of the secondary endpoint}
}
\value{
the correlation matrix, Hp goes first, and Hs goes second.
}
\description{
This function generate the correlation matrix for group sequential trials with two endpoints.
}
\examples{
corrMatGenerator(tp=c(0.64,1),
    ts=c(0.25,0.49,1),
    rhops=1)
}
\author{
Jiangtao Gou
}
