% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfun_marginalPwR.R
\name{marginalPwR}
\alias{marginalPwR}
\title{Marginal Power Rate}
\usage{
marginalPwR(cvec, t, delta = 0)
}
\arguments{
\item{cvec}{a vector of critical boundaries}

\item{t}{a vector of information times}

\item{delta}{a number shows the drift paramter}
}
\value{
a number shows the marginal power (delta isn't equal to zero) or type I error (delta is zero)
}
\description{
This function computes the marginal powers.
}
\examples{
marginalPwR(c(2.218,2.218),c(0.1,0.5,1),delta=3)
marginalPwR(1.96,t=1,delta=3)
}
\author{
Jiangtao Gou
}
